Refactor server and move routes
